trading-data-pipeline/requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7  # Für PostgreSQL
schedule==1.2.0  # Für zeitgesteuerte Abfragen